chore: Change Cachix token reference in push-backend-to-ghcr workflow.
`actionlint` complains if we refer to `CACHIX_AUTH_TOKEN` as a
`secret`, as opposed to `env`, in a reusable workflow, because
`CACHIX_AUTH_TOKEN` doesn't appear as a secret in the reusable
workflow's `secrets`.

In reality, the reusable workflow works fine, because we get
`CACHIX_AUTH_TOKEN` from Vault, dynamically & from within the workflow
itself, and not as an input. However, to satisfy `actionlint`, we can
just change the `secret.` to `env.` and it will still work.