fix: Add ID to definition name node handle
ReactFlow 11.5 (or, at least, somewhere between 11.4.2 and 11.5.6 - the change doesn't seem to be documented) started requiring that if an edge specifies a handle ID to connect to, then a handle with that ID must exist on the relevant edge. Previously, when there was only one handle it would implicitly connect to it regardless.

This upstream change is to be reverted because of the breakage caused (https://github.com/wbkd/react-flow/issues/2909), but there's no harm in being more explicit now that we know the fix anyway.