fix: Make cursor styles reflect reality
This makes it clear that nodes can be clicked on, but not edges, and that only the background can be used for drag-to-pan.

One might hope that ReactFlow would handle this for us, but it doesn't go far enough, perhaps because its main intended use case involves more interactivity, whereas we essentially just use it for rendering.