feat: Better use of z-indices, & document it.
Subsequent commits will add support for the floating toolbar on the
editor canvas, which requires a bit more discipline in our use of
CSS z-index values.

This commit makes the following changes in preparation for this:

* Move classical UI elements such as nav bars & modal dialogs higher
up the stack (`z-40` and `z-50`, respectively). Note that this creates
no visual changes: we're just expanding the range now to make room for
the floating toolbar later.

* Create a simple document where we can keep track of these uses.