Demonstration of broken bounds on Draggable.
The `FloatingToolbar` should be bound by the canvas element's
boundaries, but it's not: instead, the origin of the bounding box is
shifted left by the width of the sidebar, so that the
`FloatingToolbar` can be dragged all the way to the left-hand side of
the browser window (overlapping the sidebar), and not all the way to
the right-hand side of the canvas (offset by the width the of the
sidebar).

The same thing happens if `offsetParent` is left as the default and we
set `bounds="parent"`.