fix: Set initial floating toolbar position to upper right of canvas.
To do this, we need to read the dimensions of the canvas `div`
element. There are many, many ways to do this, but none of them are
built-in to React, as such. This Stack Overflow thread is very
helpful (perhaps overly so):

https://stackoverflow.com/questions/43817118/how-to-get-the-width-of-a-react-element

IMO, the two most promising solutions mentioned there were to use the
`use-resize-observer` package:

https://stackoverflow.com/a/71918509

and to use a React 18 technique that is apparently a more modern
approach than using `useEffect` in situations like this one:

https://stackoverflow.com/a/75101934

I settled for the latter, because the amount of code needed to
implement is very small. However, if in the future we run into
problems with browser compatibility or needing more features, we might
want to look into `use-resize-observer`.

For an explanation of why the dimensions object is string-ified, see
the Stack Overflow referenced answer. Basically, it's due to `===` not
working as you'd expect, and hopefully will be fixed in later versions
of React so that it's not necessary.

Note that per the Stack Overflow terms of service, the code in the
answer we've used is licensed under CC BY-SA 4.0. See:

https://stackoverflow.com/help/licensing
https://creativecommons.org/licenses/by-sa/4.0/
https://creativecommons.org/licenses/by-sa/4.0/legalcode

Per the terms of the license, we must document any changes we made to
the original implementation. The changes are:

* Replace `export { useDimensions }` in the original with an inline
export.

* Reformat the code using our Prettier settings.