refactor: Move `nodeTypes` check in to `ReactFlowSafe`
This is a bit more principled than the ad-hoc top-level `assertType` we were previously using. It also properly checks both the different node types used by `TreeReactFlow` and `TreeReactFlowOne` (`PrimerNodeWithDef` vs. `PrimerNode`).

As a bonus, we also now get much better type errors when `nodeTypes` doesn't match the expected shape, whether that be by failing to provide a handler for a given node type, or by requiring extra data which isn't in the declared type. With `assertType`, all we saw was success or failure, with no context.