feat: add an `onSubmit` callback to `SearchBar`
Also, we thread it all the way through the component stack, up to the
`ChooseSession` component, which is currently the only top-level
component that uses it.

Limitations of this implementation:

* It only does some generic form validation, namely stating that the
search bar value must be a string. It also trims the provided string
before submitting it.

* `ChooseSession` currently just provides a console log callback, and
doesn't do anything with it.

* There's no support here for "live search". The student must press
`Enter` or find some other mechanism for submitting the contents of
the search bar form.

* Most (all?) browsers will present a cross (`x`) button on any
`<input>` whose `type="search"`, which is the case here. The cross
button will clear the form, but unfortunately, there's no nice
built-in way to detect this specific scenario in `react-hook-form`.
This means that, as committed here, this form will not submit an empty
search string when the student clears the form with the cross button;
they must still submit the form manually after clearing it.

Clearly this latter behavior is not what we want, but the
implementation is left for subsequent commits. There are numerous ways
to do this, and I want to separate the straightforward implementation
of `onSubmit` here from the more opinionated/bespoke work required to
implement the desired clearing behavior.

Signed-off-by: Drew Hess <src@drewhess.com>