feat: add an `onChange` callback to `SearchBar`
This new callback is in addition to the previous `onSubmit` handler,
which is unchanged.

This addresses the limitation of the previous commit such that now, if
the student presses the cross (`x`) button on the search bar input
form, the caller will be notified immediately via this `onChange`
callback.

`onChange` also makes it possible to implement "live search," if
desired.

Note that you can't have your cake and eat it, too: if you want to
know when the search bar input form is cleared, then you'll be
informed of *every* change. There's no way to only be notified when
the form is cleared. (You can, of course, just ignore any `onChange`
values returned except for "", if you want the clearing behavior but
otherwise only want to know when something's been submitted.)

Signed-off-by: Drew Hess <src@drewhess.com>