fix: allow cross-origin responses to set cookies
I believe we need this setting so that the backend API can set
authentication cookies for its domain, and for the browser to include
those cookies in requests sent to the backend, for scenarios where the
frontend is making cross-origin requests to the backend. See:

https://developer.mozilla.org/en-US/docs/Web/API/XMLHttpRequest/withCredentials

This will hopefully fix the issue we're currently experiencing where,
even after authenticating to the backend API, the browser isn't
sending (or even setting) cookies for that domain when it makes
`XMLHTTPRequest`s to the backend.

Signed-off-by: Drew Hess <src@drewhess.com>