fix: filter out spurious `onChangeSearch` events that broke paging
For various reasons, the search bar component may trigger `onChange`
even when the contents of the search bar input form haven't changed;
e.g., when the page is redrawn. This broke paging through the session
list, because clicking on the "Next" page button causes the search bar
to call the `onChange` callback, which of course resets the current
session page to page 1.

Now we filter `onChange` events in the sessions page so that we only
set the session page to page 1 when the search bar input form contents
have actually changed. This filters out the spurious events.

Note that I'd prefer to have done this in the `SearchBar` component,
as we may need this functionality if/when we develop new components
that use the `SearchBar` component. However, I don't think the current
design of the `SearchBar` is amenable to this, as it's not clear to me
how to compare the current `searchTerm` value to the new value in
`react-hook-form`'s `useWatch` hook. In any case, this is left for
a future investigation.

Fixes https://github.com/hackworthltd/primer-app/issues/896

Signed-off-by: Drew Hess <src@drewhess.com>