ci: add a DCO check
This is required for third-party contributions to our Primer projects,
including this one.

Note that this workflow triggers on merge queue entries, even though
the Probot app upon which it's based doesn't have explicit support for
that. However, it may not matter in practice, as presumably any PR
that enters the merge queue will have already passed the DCO check in
the first place.

For now, this check isn't a required job, only advisory, so it won't
block anything from merging, in any case.

Signed-off-by: Drew Hess <src@drewhess.com>