feat: set `fitView` on all `ReactFlow` trees
Per the documentation, this will "fit the view once to make all nodes
visible" such that the first time the tree is rendered, React Flow
will try to fit all nodes on the canvas. This seems like something
we'd naturally want all the time, so we don't bother making it a
settable property of our React Flow wrapper type.

The zoom-to-fit functionality in React Flow is bounded by the
component's `minZoom` property, which I've tweaked here to be
hopefully small enough that it'll accommodate all reasonable programs.
It certainly zooms out far enough such that it's not possible to make
out details of trees anymore (e.g., you can't read node labels), which
is probably good enough for now, but eventually we should make an
effort to set this based on the program's bounding box.

One issue with this implementation is that it will work properly the
first time you open the live eval tree view and select an expression
to evaluate, but it doesn't re-fit the view when you select a
different expression. Presumably this is working as it's supposed
to (we seem to be re-using the same `ReactFlow` component each time we
change the expression to evaluate), and this will need some fixing,
but it's sufficient for now.

Signed-off-by: Drew Hess <src@drewhess.com>