feat: add a zoom bar to `TreeReactFlow`
Note that this component needs to be rendered inside a `ReactFlow`
component so that it can make use of `useReactFlow`, so it's not
standalone like many of our other UI components. Also, we need to
render the `Background` inside a `ReactFlow` component now, as well,
or else the background doesn't render properly.

This component clashes visually with the `TreeReactFlow` that is
rendered in the live eval window, but it's probably not worth fixing
this, as the live eval window is likely to change quite a bit in the
very near future.

Signed-off-by: Drew Hess <src@drewhess.com>