feat: scroll to def when its name is clicked in the sidebar
This approach is a bit hacky, but we couldn't think of a better way to
do this without wrapping large chunks of `App` in a
`ReactFlowProvider`, which we weren't crazy about. (However, that
approach may become less objectionable/problematic once we've broken
the live eval and definition list out of the sidebar.)

We'd prefer to pass the def's `GlobalName` to the `TreeReactFlow`
callback, but this would require quite a bit of reworking of the
sidebar's definition list. As that component is probably not long for
this world, we'll defer this work until later when we figure out
exactly where the definition list will go.

Limitations of this current implementation:

* It doesn't change the zoom level and technically only focuses on the
definition's root node, without trying to fit its entire tree on the
screen. We should probably support this, perhaps through some sort of
double-click/double-touch mechanism.

* It only works for definitions the student has defined in the current
module. It doesn't work for imported definitions, and it doesn't work
for typedefs, as those aren't yet rendered on the canvas. However,
this feature should be relatively easy to adapt to student-defined
typedefs once they're represented as trees on the canvas.

* It's not yet implemented in the live eval view. It may be useful for
fitting the tree on the screen when we switch which definition is
being evaluated, but there may be easier/better ways to do this that
don't involve the convoluted callback mechanism we've had to implement
here, because all of the controls we need are in a monolithic
component. We'll explore this in subsequent work, in any case.

Signed-off-by: Drew Hess <src@drewhess.com>