fix: use `h-full` and `overflow-hidden` on `App` page
On the `App` page, we want to ensure that:

* Each vertical section is the full height of its container, which is
the whole page, for now. (This will change soon when we add a search
bar.)

* The sidebar and action panel's contents are clipped to the container
bounds. Not that using `overflow-scroll` is not what we want here,
because each panel is responsible for deciding what should scroll and
what shouldn't. (For example, we want to pin the tab headers in the
sidebar, and will soon want to pin the input box and "Cancel" button
when prompting the student for a variable name.)

(The diff of this commit looks large, but it's almost entirely due to
indentation. This diff is probably best viewed with whitespace
ignored.)

Signed-off-by: Drew Hess <src@drewhess.com>