style: several action panel styling improvements
Specifically:

* Change the title to match the context: "Action" for actions, "Enter
a value" for a free prompt, and "Select an option" for a list of
options. (These should be made even more context-specific; e.g.,
"Enter a name" for variable prompts, vs. "Enter a value" for primitive
literal prompts.)

* The "Cancel" button is now pinned to the bottom of the screen when
the list of options overflows the container, so that it's always
handy.

Signed-off-by: Drew Hess <src@drewhess.com>