chore: replace deprecated Dicebear avatar packages
The packages we were using are no longer maintained. The replacement
packages have a slightly different API, and have dropped support for
the `jdenticon` style, unfortunately. We could add this back by using
the `jdenticon` package directly, but it would complicate the `Avatar`
type and isn't worth the time investment, for now.

Signed-off-by: Drew Hess <src@drewhess.com>