refactor: use `CreateDefModal` for both types and terms
We don't need to support value constructors in the typedef modal since
we do that via the structure editor now.

As a bonus, we now check both definitions and typedefs for duplicate
names in the current module. These checks occur in their respective
namespaces; note that this is a rare occasion where the frontend
depends on the language's semantics.

Signed-off-by: Drew Hess <src@drewhess.com>