fix: make `FloatingToolbar` only draggable via handle
Previously, the `FloatingToolbar` was draggable so long as the pointer
wasn't hovering over a button, but now that the floating toolbar
contains other controls -- and especially the `EvalFull` view, which
has its own drag interaction -- we need to be more selective.

The `@neodrag/react` component we use supports either selecting
specific elements which trigger a drag, or selecting specific elements
which will not drag (in which case any other elements will trigger a
drag). The latter behavior would be probably be preferable, and that's
what we used to do when the floating toolbar only contained buttons,
but doing this as we add additional components is tricky.

For example, consider the new `EvalFull` component. It consists of
`SelectMenu`, and an React Flow window. The `SelectMenu` itself
consists of a label element and a dropdown menu. We probably only want
to trigger a `@neodrag/react` drag when the label of the `SelectMenu`
is dragged, but not for any other element. But this would require the
ability to reach into the `SelectMenu` and add the
`neodrag-react-handle` CSS selector we're using for this feature,
but *only* to the menu label.

We could do this by adding `className` support to the various elements
of the `SelectMenu`, but that seems overly specific. For now, we
settle for just enabling drag on the `FloatingToolbar`'s handle
element in the upper-left corner.

Signed-off-by: Drew Hess <src@drewhess.com>