fix: add `title`s to nodes in React Flow trees
The `title` provides tooltips on hover, better a11y, and shows full
names when they need to be truncated to fit a node's dimensions.

Note that we place the `title` property on the overall node `<div>`,
rather than the label's `<div>`, to increase the area over which the
tooltip will be active.

Signed-off-by: Drew Hess <src@drewhess.com>