fix: truncate names on action panel buttons
This fixes the issue where very long names caused the button size to
grow unbounded. We prioritize uniform button sizes over displaying
extremely verbose names.

Note that we don't add a `title` property to these buttons like we do
for React Flow nodes, because tooltips don't seem as useful for
buttons. For one thing, action panel buttons are much wider than tree
nodes and less likely to truncate names, and even when they do, much
more of the name will be visible as compared to a tree node. Also,
tooltips are useful for tree nodes even when the node's name isn't
truncated, because the student may have zoomed out too far to make out
the node name. This argument doesn't apply to action panel buttons.

(Note that Safari will automatically add a tooltip whenever the
button's label is truncated anyway, so long as there's no `title`
property, which is a nice touch. However, this is a Safari-only
feature and I don't know of a way to emulate it in other browsers.)

Signed-off-by: Drew Hess <src@drewhess.com>