fix: truncate definition names in `Sidebar`
This fixes the issue where very long names caused list items in the
sidebar to wrap. We prioritize the uniform item height over displaying
extremely verbose names.

Note that the ellipsis doesn't show up when names are truncated in the
sidebar. I'm not sure why this happens, but I'm not too worried about
it, as the sidebar is likely not long for this world, anyway.

Also note that we don't add a `title` property to these items like we do
for React Flow nodes, for reasons similar to those described in a
previous commit where we justified not adding them for action panel
buttons. (That same commit explained how Safari will automatically add
a tooltip for truncated button labels. That does not appear to be the
case here, for reasons I don't understand.)

Signed-off-by: Drew Hess <src@drewhess.com>