refactor: break up the toolbar and live eval window
The standalone live eval component is now called `PictureInPicture`,
as we'll use it for additional visualizations in the future (e.g.,
displaying the type of the selected node). It remains draggable.

The standalone toolbar is now called `Toolbar` and is basically the
original `FloatingToolbar` implementation. Here I've just named it
`Toolbar` in anticipation of it not being draggable in the near
future.

Signed-off-by: Drew Hess <src@drewhess.com>