style: move undo+ toolbar, make it fixed
The undo/redo/level/view toolbar ("toolbar") no longer floats, and is
instead fixed in the lower right corner. We make this change because
we don't want to require the student to manage multiple floating
windows -- we don't want to recreate a window system in the canvas.

This needs some styling tweaks and is slightly out of sync with the
zoom bar. We should reconcile them, but this is left for future work.

Signed-off-by: Drew Hess <src@drewhess.com>