style: move undo button to far right of `Toolbar`
We assume undo will be the most used of these buttons, so if we ever
manage to remove the `ActionPanel`, this will put undo close to the
lower right corner of the screen, taking advantage of Fitt's law.

Also, we put the level button on the far left, because it will be the
least used button, and therefore it could potentially be hidden behind
a fold-out chevron at the left edge of the toolbar.

Signed-off-by: Drew Hess <src@drewhess.com>