feat: add support for wildcard patterns
Here we've decided to render wildcard patterns specially, without
using any of the value constructor UI, and using a light-hearted shrug
emoji (🤷🏽‍♀️). The styling is subject to change based on feedback, but
in any case, in order to avoid confusion, we were keen to avoid
anything that resembles an empty hole.

We also didn't feel particularly compelled to use the traditional `*`
or `_` characters, though we would probably choose the former (or more
likely some Unicode variant of it) if no obviously better choice were
available.

(Note that the tooltip on hover reads `PatternWildcard`, which is
perfect.)

Signed-off-by: Drew Hess <src@drewhess.com>