feat: add `FitViewParams` to `TreeReactFlow`
This type is a subset of `ReactFlow`'s `FitViewOptions`, and gives us
the ability to set `TreeReactFlow`'s initial `fitViewOptions` property
to match its `ZoomBar`'s behavior.

With this change, views in the evaluation PIP are more consistent:
when you select a definition to evaluate, what you initially see is
exactly what you get when you press the "fit" button.

Signed-off-by: Drew Hess <src@drewhess.com>