fix: add `children` support to `TreeReactFlow`
With this change, we can render React components inside the start and
end of a `TreeReactFlow` component. Among other things, this removes a
bunch of hacks we needed to asynchronously determine the size of the
canvas and re-render components when it changes.

Now that we can parent one `TreeReactFlow` inside another, we need to
add a `ReactFlowProvider` around each use of `ReactFlow` (aka
`ReactFlowSafe`) in order for them to coexist. At the moment, this is
done internally in the `TreeReactFlow` implementation, but we may want
to move this outside and make it the caller's responsibility, if we
need more flexibility later (e.g., to reference `ReactFlow` internals
from other UI widgets).

Signed-off-by: Drew Hess <src@drewhess.com>