feat: add `children` support to `TreeReactFlow` and `TreeReactFlowOne`
With this change, we can parent React components to `TreeReactFlow`
and `TreeReactFlowOne`. Among other things, this removes a bunch of
hacks we needed to asynchronously determine the size of the canvas
when rendering the `PictureInPicture` and `Toolbar` components, which
overlay it.

Now that we can parent one `TreeReactFlow` inside another, we need to
add a `ReactFlowProvider` around each nested use of React Flow in
order for them to coexist. This is the responsibility of the caller to
get right, the advantage being that we can now get & use internal
React Flow state from pretty much anywhere.

Signed-off-by: Drew Hess <src@drewhess.com>