fix: set a `key` on `Evaluated` trees
This fixes a subtle but potentially important UX issue in the
grandparent commit. Recall that the grandparent commit re-fits trees
in the `EvalFull` view when their nodes change. That feature works
great for most use cases, but as the following example demonstrates,
there's potential for confusion:

1. The student chooses a definition whose value is, say, a list `[1,
2, 3]`.
2. The student pans and zooms the expression in the `EvalFull`
component in order to examine it.
3. The student switches the evaluation selection to a different
definition whose value is also a list `[1, 2, 3]`.

In this case, because the nodes haven't changed, the view won't
change, either; i.e., the value of the newly-selected definition won't
be zoomed to the default zoom setting, nor re-fitted in the viewport.
This could be interpreted as a bug.

With the simple change in this commit, we also force React to
re-render the evaluated tree whenever the selected definition changes,
which solves the issue with the above scenario.

Note that we still need the grandparent commit's functionality,
because that will cause the `EvalFull` view to re-fit when the
evaluated definition's value changes, which this commit does not do.

(One hopes that React is smart enough not to render trees twice now
when the student changes the selected definition and their values are
different, but I haven't verified that.)

Signed-off-by: Drew Hess <src@drewhess.com>