feat: add selection type info to `PictureInPicture`
We add tabs to the PIP, and add the new type info panel as a new tab
in addition to the existing eval view.

The styling here needs a lot of work, but this is mostly functional,
except for 2 known issues:

* Upon initial load, even if there's a selection, it doesn't show up
in the info panel for some reason.

* The types of holes are shown as a hole themselves. This is due to a
backend bug that we haven't figured out yet. All other types I've
tried seem to be correct.

Signed-off-by: Drew Hess <src@drewhess.com>