fix: remember the current eval def when the eval tab is deactivated
Now that the `EvalFull` component isn't always rendered (i.e., if the
selection info tab is selected in the PIP), it can't persist the
currently selected definition. However, the `Edit` component knows it,
so we pass it down to `EvalFull` via a new property.

(The whole eval roundtrip from `EvalFull` back up to `Edit` is pretty
convoluted, and we should figure out a better way to do this.)

Signed-off-by: Drew Hess <src@drewhess.com>