feat: add checkbox for importing Prelude on new program
This needs some styling fixes and could use some explanatory text, at
least at beginner level, but it'll do for now.

Ideally this would be implemented with Headless UI's `Switch`
component, but that is surprisingly tricky to get to work with `zod`,
so I've just resorted to a standard checkbox here, despite those being
unusual on touch interfaces. I've made the checkbox a bit comically
large in order to make it an easier touch target.

Also, the we should obey the student's previous preference for
importing the Prelude, but we need an `onChange` callback to make that
work, and I'm short on time. Left for later improvements.

Signed-off-by: Drew Hess <src@drewhess.com>