style: restyle definition & typedef nodes
Mainly we bring the border styling in line with other nodes and make
the label font less loud. I think we already draw enough attention to
these special nodes due to their size, but here we also use a slightly
"electric" border color and also increase the font size and weight a
bit vs. normal nodes.

We've also increased contrast by ditching the grey background. (We
would like to move towards a grey canvas background, anyway, so the
old background color would have clashed.)

One drawback to the new style is that definition names are harder to
see when zoomed out, but we have other ways to locate
definitions (e.g., clicking on their names in the sidebar), and we
have tooltips on hover, as well. (One benefit to making the font size
smaller is that definition names can be longer before they're
truncated.)

Signed-off-by: Drew Hess <src@drewhess.com>