feat: Use node shapes to distinguish between terms, types and kinds
The choices of shape are somewhat arbitrary, but there is a little logic to it: types are rectangular because they're in a sense more "rigid", and kinds are drawn as diamonds since each node currently only has a single character.

We considered using `clip-path`, instead of rotation, to create the shape for kinds, but this makes drawing borders difficult as they don't follow the path. The usual trick would be to add a dummy background to create a manual border, but this quickly gets messy when we don't want it for the type or term case.

Signed-off-by: George Thomas <georgefsthomas@gmail.com>