feat: Use colours for all node flavors
Well, almost all. We still use black for type annotations due to running out of colours in our existing palette.

We try to colour type nodes analogously to corresponding term nodes. This makes more sense now that we use different shapes for terms and types, but it probably requires a little more thought. For example, it may seem odd that we style term `Lam` and type `LAM` differently, since we choose instead to highlight their analogies with `TFun` and `TForall` directly.

We should consider refactors to make this sort of change less laborious, though it's hard to say how to abstract colours properly when tailwind discourages concatenating class strings.

Signed-off-by: George Thomas <georgefsthomas@gmail.com>
Signed-off-by: Drew Hess <src@drewhess.com>