style: render misfit label as ⚠️
Note that I tried a few emoticons for empty holes, as well, but they
didn't work out. One looked like an eye, and the other (a golf flag in
a hole) has some issues, such as implying that the hole is filled, and
I'm a bit worried about the golf idiom not being universal. Let's
stick with `?` for now.

Signed-off-by: Drew Hess <src@drewhess.com>