feat: Use left arrows for applications
This covers _all_ sorts of application: term-to-term, term-to-type, and type-to-type. We were already using `@` for two different concepts, so we may as well unify all three. And `$` is a weird Haskell-ism that isn't really used elsewhere. We can now easily tell which sort of application is in use by the shape of children.

Signed-off-by: George Thomas <georgefsthomas@gmail.com>