fix: remove `Sidebar` to free up more horizontal space
We do the following:

* Remove the `Sidebar` altogether.

* Move the "Add new definition" and "Add new type" pseudo-actions to
the `ActionPanel`.

The consequences are:

1. Pro: we gain a lot of horizontal room, making smartphone form
factors more viable, and small laptop screens much more comfortable
for editing.

2. Con: we conflate global actions with contextual actions in the
`ActionPanel`.

3. Con: we lose the ability to see a list of definitions & types, and to
scroll to them upon clicking their names.

We feel that the benefits of 1 outweigh the drawbacks of 2 and 3. We
can get back the functionality described in point 3 fairly easily, and
less obtrusively than the sidebar did it. (That work is out of scope
for this commit.)

On point 2, I would like to consider sooner than later whether even
the (static) action panel is necessary. A new design will give us a
chance to separate the global actions from the contextual ones again,
but after playing around with alternatives to these style changes, I
feel like we're stuck putting the global actions in the action panel
with the current design: there's simply no other good place for them.

Signed-off-by: Drew Hess <src@drewhess.com>