fix: scroll to new definition when created
Note that this is implemented via a pretty nasty hack: we `setTimeout`
for 100 ms before scrolling to the definition, because immediately
upon return from the successful API call, React Flow won't have had
time yet to render it.

We'll need to set up some kind of callback system from our
`TreeReactFlow` component to do this properly, but this will do for
now.

Signed-off-by: Drew Hess <src@drewhess.com>