fix: show global actions when there is no selection
When I removed the sidebar in
e3428ec1811fc94336d8fa0f416f896c7b215cd1, I didn't consider what
happens when there's no selection: we don't render the action panel,
but that's where the "add new def" and "add new type" buttons are now!

We now render just those 2 actions in the action panel when there's no
selection. This has the nice side effect of getting rid of the not
very helpful and hastily designed message that we used to display when
this happens.

Note that the "fixes" are more like hacky workarounds. The action
panel wasn't designed to be shown without a selection, so we pass some
console logging `on*` handlers to catch any problems that might arise
due to this change. We should refactor things to properly support this
scenario, though.

Signed-off-by: Drew Hess <src@drewhess.com>