fix: always render the session page nav bar
The previous commit improved matters on the session page by handling
`isError` and `isLoading` conditions, but has annoying behavior when
the student is typing a program name into the search bar and we're
doing live search, because the sessions page was rendered
all-or-nothing.

Now we always show the nav bar. This means we no longer have a nice
single component for the sessions page (i.e., the `SessionsPage`
component is no more), but `ChooseSession` is now more or less the
same thing, only with better error and loading behavior. I haven't
bothered to add a Storybook story for it, but it would be easyenough
to add later if we need it.

Signed-off-by: Drew Hess <src@drewhess.com>