fix: Avoid anti-aliasing issues in "inline"-style nodes
We had an issue where there were sometimes small visible gaps between the node label and border. These gaps would flicker while zooming in and out. This was seen across multiple browsers.

We fix this by adding extra padding around the label so that it intersects the border, rather than meeting it precisely.

There is a side effect here, which is that labels in term-level nodes are now curved on both sides (I wouldn't have known how to implement this otherwise). But this is arguably an improvement anyway.

Signed-off-by: George Thomas <georgefsthomas@gmail.com>