feat: Display and allow editing of typedef parameter kinds
This bumps the Primer pin, and adds support for the new actions and selections, as well as displaying the kinds as part of the typedef tree. The way these are displayed is, for now, not especially intuitive, but it's the best we can do without redesigning the typedef UI or adding further annotations.

Signed-off-by: George Thomas <georgefsthomas@gmail.com>