feat: Add devtools menu
This retains the ReactQuery devtools panel, but also has an extra area to the right (unused right now) in which we can add further controls. We can use this for gating features which are experimental or which help with debugging.

In production builds, this can be triggered as before, by running `window.toggleDevtools()` in a browser console.

Signed-off-by: George Thomas <georgefsthomas@gmail.com>