feat: Add some checkbox controls to devtools menu
One allows us to display node IDs, which can be useful for debugging. Note that, for now at least, we only show these for the "primer", "primer-simple" and "primer-box" nodes. For others, it's less obvious where to put them, since we don't have labels as such. And they'd be less useful anyway, since the mentioned node types happen to be the ones whose IDs are backend-generated.

The other allows us to enable the expert-mode syntax-label-hiding feature.

Signed-off-by: George Thomas <georgefsthomas@gmail.com>